+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- fixed: Prevent fatal error reporting for missing txlist json file
+
 ## 4.14.0 (2024-07-19)
 
 - added: Added Amoy Polygon testnet