+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - changed: (Zcash) Updated address explorer
+- changed: (Tron) Special case the `usdt-trc20` uri prefix
 - fixed: Prevent fatal error reporting for missing txlist json file
 
 ## 4.14.0 (2024-07-19)